remove Inbox module
